backend/requirements.txt
appdirs==1.4.4
arrow==1.2.3
asgiref==3.5.2
azure-cognitiveservices-speech==1.24.2
beautifulsoup4==4.11.1
binaryornot==0.4.4
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
coverage==5.5
cryptography==41.0.1
cssselect==1.2.0
defusedxml==0.7.1
deprecation==2.1.0
Django==4.1.3
django-browser-reload==1.6.0
django-cors-headers==3.13.0
django-json==0.27
django-taggit==3.0.0
django-tailwind==3.4.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
eyed3==0.9.7
fake-useragent==1.1.1
filetype==1.2.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.1.0
importlib-resources==5.10.1
Jinja2==3.1.2
jinja2-time==0.2.0
lxml==4.9.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==23.1
parse==1.19.0
Pillow==9.5.0
pycparser==2.21
pyee==8.2.2
PyJWT==2.7.0
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-slugify==7.0.0
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-html==0.10.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
soupsieve==2.3.2.post1
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
urllib3==1.26.12
w3lib==2.1.0
websockets==10.4
zipp==3.11.0
